move QubitIdentifierType fully to types
