+++ src/braket/experimental/autoqasm/instructions/instructions.py
@@ -17,8 +17,9 @@
 from typing import Any
 
 from braket.experimental.autoqasm import program as aq_program
+from braket.experimental.autoqasm.types import QubitIdentifierType
 
-from .qubits import QubitIdentifierType, _qubit
+from .qubits import _qubit
 
 
 def _qubit_instruction(
